<bugfix>MultiHeads configuration was not working on 2.0 </bugfix>
The problem was introduced since we started VisTrails with the Spreadsheet hidden and the setup code for multiple displays and window maximization was not performed anymore.
The fix was to run the setupCode the first time the Spreadsheet window is shown.

--

packages/spreadsheet/spreadsheet_window.py
class SpreadsheetWindow:
 __init__():
    added new member self.shownConfig

configShow()
    perform setup code when self.shownConfig is False
    set self.shownConfig = True

packages/spreadsheet/basic_widgets.py
In SpreadsheetCell.displayAndWait()
    calling spreadsheetWindow.configShow() instead of spreadsheetWindow.show()