<bugfix>Fixed issue with forced versions on modules</bugfix>
In 2.0, modules from workflows that do not have package version information (workflows produced in earlier versions of VisTrails) must now have version information and are upgraded to do so.  There was an issue with the check for version information (checking == '' and not == None) which is now fixed.

core/vistrail/pipeline.py:
  - Pipeline.ensure_modules_are_on_registry: check if evaluating
    module.version as a boolean is False rather than if
    module.version == '', this captures None and ''