<bugfix>Fix for remove_connection issues</bugfix> <ticket>#522</ticket>
The issue seems to be rooted in the fact that the module configuration
was setting a hidden variable with the ports that had been set on
modules like PythonSource boxes, and that was used during
recreate_module calls.  Because property windows are updated when
events fire, this configuration code was being called when switching
versions instead of only when the module is being configured.  This
caused the connections to be removing the connections.  This fix
changes things so that dependent connections are stored on the
modules, mirroring the pipeline graph instead of relying on it so that
the current state of the items can be more easily reconciled with the
pipeline graph.  In addition, delete connections are now done first
instead of after recreate_module calls which minimizes the connection
issues.

gui/module_configuration.py:
  - QModuleConfiguration.updateModule: remove _old_connection_ids setting

gui/pipeline_view.py:
  - QGraphicsPortItem.removeVars: use new style
    dependingConnectionItems, remove unused code
  - QGraphicsModuleItem.__init__: default connectionItems
  - QGraphicsModuleItem.moduleHasChanged: new method, refactor code
    that used to be in setupScene
  - QGraphicsModuleItem.addConnectionItem: new method
  - QGraphicsModuleItem.removeConnectionItem: new method
  - QGraphicsModuleItem.dependingConnectionItems: now returns dict
  - QGraphicsModuleItem.dependingConectionsItemsWithDir: new method
  - QGraphicsModuleItem.itemChange: use new style dependingConnectionItems
  - QPipelineScene.remove_connection: update module connection items
  - QPipelineScene.recreate_module: don't worry about old dependent
    connections any more, we delete old connections first!
  - QPipelineScene.module_text_has_changed: moved to moduleHasChanged method
  - QPipelineScene.setupScene: delete old connections first, refactor
    out module change checking code
  - QPiplelineScene.delete_selected_items: use new style
    dependingConnectionItems
  - QPipelineScene.get_selected_item_ids: use new style
    dependingConnectionItems