<bugfix>Fix issue with configure item grabbing mouse.</bugfix>
There was an annoying issue where clicking on the contextual triangle menu would send Qt into a state where clicking on other modules would bring up the contextual menu for the originally selected module.  This seems to be due to this item becoming the mouse grabber item.  Now, we accept the event and call ungrabMouse which seems to fix the issue.

gui/pipeline_view.py:
  - QGraphicsConfigureItem.mousePressEvent: accept the event and
    ungrab the mouse