<bugfix>Fix logic in identifier checks in package manager</bugfix>
Not sure what was going on here as it is old code, but it looks like
the indentation may be off.  It seems like we want to return the the
package when we find the correct identifier, but only if it loads
properly.  This has been changed.  Also added a catch-all if something
goes wrong when loading (set_properties can generate exceptions).

core/packagemanager.py:
  - PackageManger.identifier_is_available: change indentation to
    reflect logic and catch other types of exceptions from packages.