<ticket>#534</ticket>
--
Problem: Mashups hang with auto-update on
When auto-update is on, an execution is triggered every time the user changes a parameter. This will cause VisTrails to hang if an execution is already happening. The solution was to run only when a workflow was not  already running.
Also disabling the update button until the execution is finished (when auto-update is off)

---

gui/mashups/mashup_app.py
   class QMashupAppMainWindow:
     new method
        updateButtonClick()
     __init__()
        Added is_executing control variable
        Connecting click event of updateButton to updateButtonClick
     updateCells()
        updating execution status before and after calling self.runAndGetCellEvents()
     widget_changed()
        Only calling updateCells() when auto-update is on and it is not executing