<feature>AutoConnect: when dragging a new module into a workflow, try to connect it</feature>
The AutoConnect feature will try to connect a new module with the existing modules in a workflow based on the location and compatibility of the added module.  The proposed connections are displayed as the user drags the new module into the workflow view.  AutoConnect will attempt to connect at most one input and one output port.  Currently, this is accomplished only on type compatiblity and location, but we could use information about already connected ports and/or developer-specified suggestions.  The feature can be turned on or off in the preferences (the "autoConnect" setting).

Right now, there is an issue with setting two connections (both input and output) as only one tooltip can only be displayed and so you see a flicker.  There is also an annoyance when multiple inputs and multiple outputs are compatible (with only one side with multiple choices, things are ok).  Specifically, if port1 and port2 are compatible with portA, portB, and portC, depending on location (and distance computations), we might choose port1--portA or port2--portB, and things may flip-flop between the two.  Here, we may choose to bias toward the first port or use a developer-specified ordering to try matches.

This feature also incorporates an image of the module being added when dragging so users can see exactly where the module they are adding will be placed.  Previously, text was shown, but it was hard to know what the module would look like once added and additional moves of that modules were often necesssary.  The image scales based on the scale of the pipeline view.

core/configuration.py:
  - default: add autoConnect setting

gui/module_palette.py:
  - QModuleTreeWidget.startDrag: new method

gui/pipeline_view.py:
  - QAbstractGraphicsPortItem.__init__: connection -> tmp_connection_item
  - QAbstractGraphicsPortItem.add_connection_event: **moved** to
    QPipelineScene.addConnectionFromTmp
  - QAbstractGraphicsPortItem.mouseReleaseEvent: use new methods to
    control adding and selection
  - QAbstractGraphicsPortItem.mouseMoveEvent: changed to use
    QGraphicsTmpConnItem and common methods
  - QAbstractGraphicsPortItem.findModuleUnder: **moved** to QPipelineScene
  - QAbstractGraphicsPortItem.findSnappedPort: **deprecated**, use
    QPipelineScene.findPortMatch
  - QGraphicsTmpConnItem: new class
  - QGraphicsModuleItem.paintToPixmap: new method
  - QGraphicsModuleItem.getClosestPort: new method
  - QGraphicsModuleItem.getDestPort: use getClosestPort
  - QGraphicsModuleItem.getSourcePort: use getClosestPort
  - QPipelineScene.__init__: set tmp_module_item, tmp_input_conn,
    tmp_output_conn
  - QPipelineScene.findModuleUnder: new, moved from
    QAbstractGraphicsPortItem
  - QPipelineScene.findModulesNear: new method
  - QPipelineScene.findPortsNear: new method
  - QPipelineScene.findPortMatch: new method
  - QPipelineScene.updateTmpConnection: new method
  - QPipelineScene.updateTmpInputConnection: new method
  - QPipelineScene.udpateTmpOutputConnection: new method
  - QPipelineScene.dragEnterEvent: support autoConnect
  - QPipelineScene.dragMoveEvent: support autoConnect and rewrite
    global variables code to use findPortMatch
  - QPipelineScene.dragLeaveEvent: new method
  - QPipelineScene.createConnectionFromTmp: new method
  - QPipelineScene.addConnectionFromTmp: new method, moved from
    QAbstractGraphicsPortItem.add_connection_event
  - QPipelineScene.add_module_event: support autoConnect
  - QPipelineScene.add_tmp_module: new method, support autoConnect
  - QPipelineScene.delete_tmp_module: new method
  - QPipelineView.paintModuleToPixmap: new method