<bugfix>Fix issue with caching group modules</bugfix> <ticket>#558</ticket> <feature>Allow modules in subpipelines to be cached</feature> <ticket>#559</ticket>
Wrote a new signature function for the Group module which takes into account both the subpipeline and the input connections.  This allows groups to be properly cached (Ticket #558) and to allow interior pieces of groups to be cached as well (Ticket #559).  Basically, we write the upstream signature of the ports in a group to the signature of the interior InputPort module (we write it to _input_port_signature on that module).  Then, we recompute the signature of the subpipeline (a concatenation of the signatures of all of the sinks) using these new signatures.

Note that there is still a difference between the group signature and the signature of the subpipeline the group replaced.  Thus, running a pipeline, grouping a region, and running the new pipeline will not use the old cache as desired.  This is mostly due to the fact that groups define InputPort and OutputPort modules that become part of the Group signature so the signatures inside the group become different than those outside.

The is_cacheable method is still written so that any group with a non-cacheable module will be non-cacheable, but I believe we can remove this so that even groups with non-cacheable modules can be partially cacheable.  Also, we should see if these changes affect control flow modules.  This should also work for nested groups.

core/modules/sub_module.py:
  - random_signature: moved out of initialize
  - input_port_signature: new method
  - group_signature: new method, computes the signature of the
    interior of the group with the external upstream connections and
    also allows caching in the subpipelines
  - initalize: change InputPort and Group signatures