<bugfix>Fix undo/redo in pipeline view</bugfix> <ticket>#517</ticket>
Each view has its own actions that are enabled/disabled only when that
view is active.  Thus, the checks in the version view do not happen
when we're in a pipeline view.

We move the redo_stack and most of the implementation to the controller instead of having it in the vistrail view.

Also, change the change_selected_version call in vistrail view to pass True to the check for errors rather than the value of by_click.

core/vistrail/controller.py:
  - VistrailController.__init__: set redo_stack
  - VistrailController.reset_redo_stack: new method
  - VistrailController.undo: new method, moved from gui/vistrail_view.py
  - VistrailController.redo: new method, moved from gui/vistrail_view.py:
  - VistrailController.can_undo: new method
  - VistrailController.can_redo: new method

gui/pipeline_view.py:
  - QPipelineView.set_action_links: add undo, redo
  - QPipelineView.can_redo: new method
  - QPipelineView.can_undo: new method

gui/version_view.py:
  - QVersionTreeView.__init__: remove redo_stack
  - QVersionTreeView.can_redo: use controller
  - QVersionTreeView.can_undo: use controller

gui/vistrail_view.py:
  - QVistrailView.version_selected: pass True for error reporting
    instead of the value of by_click, call reset_redo_stack
  - QVistrailView.set_pipeline_selection: remove optype parameter
  - QVistrailView.undo: most of the functionality moved to controller
  - QVistrailView.redo: most of the functionality moved to controller