<bugfix>Fix issue with connected indicator for ports</bugfix> <ticket>#569</ticket>
The connected indicator was being shown when it should not have been due to an issue with the state bookkeepping.  When a pipeline was copied, the connected counts were incremented without being reset, leading to inflated counts of the number of connections to a given port.

core/vistrail/pipeline.py:
  - Pipeline.set_defaults: reset the connected_input_ports and
    connected_output_ports dictionaries before adding to them