<bugfix>Fix issue with analogies that also require upgrades</bugfix>
Because analogies use the *actions* between two versions, and upgrades work on materialized pipelines, the resulting pipeline from applying an analogy can require an upgrade.  In this case, we should make sure that upgrade is materialized immediately.  In addition, we should reset _delayed_actions when handle_invalid_pipeline is not being used as part of change_selected_version.

gui/vistrail_controller.py:
  - VistrailController.perform_analogy: reset _delayed_actions after
    each handle_invalid_pipeline call, and flush any upgrades after
    the analogy finishes.