<bugfix>Now persistent cache works after one execution</bugfix>
Persistent cache was only working after two executions of the workflow.
This happens because the version is unknown during the first signature calculation,
as it is only included in the repository after the execution.
This is a fix, proposed by David, to solve this issue. Basically, there is now
a new function (persistent_module_hasher) that does not consider the id and version
when reading from local (when the version is not yet in the repository).

packages/persistent/init.py:
- persistent_ref_hasher: commented out
- persistent_module_hasher: new function