Started writing a Weka package: Java source parser.
The goal is to autogenerate the VisTrails Module's from the Weka JARs. However, the name of the method parameters can't be read by reflection; parsing the Java source JAR (weka-src.jar, included in the distribution) seems the only solution.
This changeset provides a working Java parser which will be used for that task.