<ticket>#445</ticket>
FIxed problem where the preview window was not always shown.

I was finally able to reproduce this problem consistently and I think I finally fixed it. The problem was when the focus was on another widget just before pressing the Peview button and that was causing a view change (resetting the view, so it would remove all the preview tabs).

--
gui/vistrails_window.py
  QVisTrailsWindow.applicationFocusChanged()
      Making sure that the Mashups Preview Tab is in the list of widgets that do not
      generate a view change