<bugfix>Add default support back in</bugfix> <ticket>#523</ticket>
Add support for default values for parameters in a way that mimics older versions of VisTrails.  There is a new solution (for 2.1) that allows parameter widgets to configure default settings themselves and enforce default values in the compute methods, but it requires some schema support.

For this solution, whenever a user expands a function that has defaults, those values are immediately written to provenance.  Thus, if a user wants to remove such parameter settings, they must delete the function.  Also, each add will add default values in as well.  Note that this is slightly different from parameters that do not have defaults as nothing is written to provenance on expansion, only when a value is set.

Also, refactor some code so that we have one place to update for
adding methods.

gui/ports_pane.py:
  - PortsList.item_clicked: call do_add_method here now
  - PortsList.do_add_method: new method, factors code in item_clicked
    an add_method, also adds default value support
  - PortsList.add_method: call do_add_method here now