<bugfix>Ensure add_*_port can override ports in superclass port lists</bugfix> <ticket>#404</ticket>
When a port in a superclass is overridden by a definition in initialize, we were overriding it again with the original superclass definition when, for example, _input_ports is inherited.  To fix this, we check that port list is defined by the subclass, otherwise, we do not add it.

Also factors auto_add_ports so that similar code for adding inputs and outputs is unified.

core/modules/module_registry.py:
  - ModuleRegistry.auto_add_ports: refactor, check if port list is
    defined on the specified class