<bugfix>Fix issues with package list in preferences</bugfix> <ticket>#584</ticket>
Move the package_added signal to the same location as the add_menu_items call so that we know the package was actually added.  Also, just let package_added do the lifting for updating the lists---remove code at the end of the enable/disable calls.

core/packagemanager.py:
  - PackageManager.late_enable_package: add package_added notification
    at the end of this method
  - PackageManager.initialize_packages: add package_added notification
    at the end of this method
  - PackageManager.availabe_package_names_list: use a set and add in
    the _package_list keys

gui/preferences.py:
  - QPackagesWidget.enable_current_package: don't update the list
    here, just use package_added
  - QPackagesWidget.disable_current_package: don't update list here,
    just use package_removed
  - QPackagesWidget.package_added: grab the
    available_package_names_list again here if the package is not found