<bugfix>Guard against empty group modules</bugfix> <ticket>#592</ticket>
This commit changes two things.  First, it adds checks to see if the workflow that a group represents is None, and if so, it will not continue to process the group.  Second, it hides the Group and Subworkflow modules in the module palette so that users do not drag those modules into the workflow; instead, they create them via the menu commands.

core/inspector.py:
  - PipelineInspector.inspect_spreadsheet_cells.find_spreadsheet_cells:
    check if the subpipeline is not None

core/modules/sub_module.py:
  - initialize: add hide_descriptor=True for Group, Subworkflow

core/vistrail/pipeline.py:
  - Pipeline.validate: only recursively validate on subpipelines when
    they are not None