Added extra.core.java_vm, a wrapper for the Java VM.
This module provides a single method to access a Java Virtual Machine, using either JPype on CPython or Jython's capabilities if we are using this interpreter.
The weka package have been updated to use this interface, meaning that it now runs without Jython as well.

The get_java_vm() returns an object which can be used to access Java packages. It uses the CLASSPATH environment variable to build the -classpath VM argument (which cannot be altered for JPype).
The get_class() function replaces Class.forName() which has some issues on JPype (cannot access the classloader from the caller, since it's JNI and not Java, and specific reflection methods).
The JavaException class is used to catch a Java Exception (WIP: we cannot catch a specific subclass).
The build_jarray() function replaces Jython's jarray.array and JPype's JArray. It provides a way to use primitive (non-wrapper) types as well.

On both CPython and Jython, the CLASSPATH environment variable must now  be set to contain weka.jar prior to execution.