Added a java_requirements package.
Similar to the GUI checks, a check_java_requirement function has been added. It is called when building the application to ensure the required JAR can be imported from Jython.
If a package is not available, the function will try to locate it, prefering the exact filename (i.e. version) that was requested, and add it to the path.