Unload dependent packages if they don't live in the system install.
Previously, all dependent packages (i.e. that are imported during the
initialization of a module) would be remembered, to be unloaded when
the VisTrails package is unloaded, provided that they were not already
loaded when imported.
Now, packages that come from the system install of Python are not
remembered to be unloaded. Packages provided with the application
should still behave the same way.
This behavior can be overriden by setting the 'force_unload_packages'
and 'force_dont_unload_packages' to a set of packages variables in a
VisTrails package's __init__.py.