<feature>Allow selected portions of the workflow to be relaid out</feature>
A few changes to the workflow relayout code including the ability to select a portion of the workflow and only layout that part.  It also recenters the layout now and incorporates any moved modules that were not selected into a single move action.  Because we are mostly dealing with the graphical positions here, I moved the code to pipeline_view. However, it still may make sense to move portions of this code down to the core level.  For example, allow add_module to pass None for x and y coordinates and choose coordinates based on a layout.

core/layout/workflow_layout.py:
  - remove commented out code

gui/pipeline_view.py:
  - QGraphicsModuleItem.__init__: default handlePositionChanges
  - QGraphicsModuleItem.setupModule: use setPos instead of deprecated
    resetMatrix/translate calls
  - QGraphicsModuleItem.itemChange: check if we are handling position changes
  - QPipelineScene.layout: moved code from VistrailController, added ability to use selected modules only, handle connection redraw internally, and use the flush_delayed_actions to generate the move (and make sure upgrades also happen)
  - QPipelineScene.selectAll: don't select ports

gui/vistrail_controller.py:
  - VistrailController.layout_workflow: ** moved ** to QPipelineScene.layout