<bugfix>Fixed issue with connecting tuple ports in the API</bugfix>
There was an issue where adding ports like (Integer, Float) using the outputs from two different modules did not work properly.  There were a variety of issues including PortSpec not existing, an incorrect construction of the tuple module, and the module returned from the add_module* calls in the api not corresponding to the actual one in the pipeline.  Also fixed __getattr__ calls to match the fact that python calls them last.

core/api.py:
  - Module.__getattr__: just handle the port specs here or raise exception
  - Module._update_func: fix issues with tuples, use api add calls, use Module for tuple
  - Package.__getattr__: just handle the registry call here
  - VisTrailsAPI.add_module: pass back the actual module in the pipeline
  - VisTrailsAPI.add_module_from_descriptor: pass back the actual module in the pipeline
  - VisTrailsAPI.add_port_spec: call add_module_port, not add_port_spec which does not exist