<feature>Added some modifications in the code to allow the execution of workflow in IPython engines</feature> <bugfix>Groups and Subworkflows can now have their execution log annotated</bugfix>
Some modifications in the code were necessary to allow the execution of a module / group in the IPython engines. core.modules.vistrails_module.Module has now an attribute (annotate_output) used to indicate that the outputs should be annotated in the log, and a method was also created to do that; this was necessary so we could get the desired output data from the engines. As we serialize the workflow to send to the engines, we need to set this attribute in the engines, after unserializing the xml. So, an annotation is added to the module, to indicate that the output, after the execution, should be stored. Some code was added to core.interpreter.CachedInterpreter.setup_pipeline to identify this annotation and set annotate_output to True. Last, in core.application, it was necessary to add an optional argument (args) for the initialization of the VisTrails application; when the VisTrails API was being executed in the engines, the CommandLineParser was getting command line arguments from IPython, and then, the initialization of the VisTrails application was failing; the solution was to pass to the initialization of the application an empty list of arguments.
Besides these modifications, a bug with annotation in core.module.sub_module was fixed. In core.log.controller.LogController.insert_module_annotations, the attribute is_group is verified so we can check if either group_exec or module_exec should be used. However, this attribute was only defined in core.vistrail.group.Group, and not in core.modules.sub_module.Group, and the latter is the one being used inside LogController. The same attribute was added to core.modules.sub_module.Group (Abstraction inherits Group, so the same fix is applied to subworkflows).

vistrails/core/application.py:
- init: added an optional argument named args
- VistrailsApplicationInterface.setupOptions and VistrailsApplicationInterface.init: added an optional argument named args
- VisTrailsCoreApplication.init: added an optional argument named args

vistrails/core/interpreter/cached.py:
- CachedInterpreter.setup_pipeline: added checking for annotation that indicates that the output should be stored

vistrails/core/modules/sub_module.py:
- Group.__init__: added attribute is_group

vistrails/core/modules/vistrails_module.py:
- Module.__init__: added attribute annotate_output
- Module.annotate_output_values: new method that annotates the output values
- Module.update: checks now annotate_output