Some initial code on building an improved matplotlib package.
The purpose here is to redesign the matplotlib package to be more user-friendly, alleviating the need for manual scripts in most places.  In addition, the spreadsheet cell is redesigned so that the buttons appear in the toolbar and the cell is free of navigation icons.  Finally, it makes use of the new enumerated value support for parameters, meaning users can pick items like markers from a drop-down list instead of consulting the matplotlib manual for the marker string shortcuts.

This is mostly accomplished by parsing docstrings in the matplotlib manual.  Ideally, this will be a mostly automated process where parsing can be done and then overridden with specific changes that need to be made (e.g. generate a registry xml automatically and then edit that---when regenerating xml, we can comparse if anything changed from the old version and try to update that---if it was overwritten, we flag it for the developer).

The current basic setup is to use the high-level pylab commands to generate plots (e.g. plot, hist, boxplot), but then attributes can be changed via calls to the artists (lines, patches).  The new MplBoxPlot module takes advantage of this to allow users to control the appearence of the box, whiskers, and outliers individually.

packages/matplotlib:
  - work in progress on a new matplotlib package