<feature>Module / Group executions from the IPython engines are added to the provenance</feature>
All the module / group executions from the IPython engines are now added to the provenance of the workflow. Besides, an annotation on these executions is made so the user can identify them as parallel executions. The schema was not changed yet to add machine information.

vistrails/core/interpreter/cached.py:
- CachedInterpreter.setup_pipeline.set_computing: new function to put the color of a module as "computing"
- CachedInterpreter.setup_pipeline.add_exec: new function to add a log execution to the current parent log execution
- CachedInterpreter.setup_pipeline.logging_obj: added set_computing and add_exec

vistrails/core/log/controller.py:
- LogController.add_exec: new method to add a log execution to either the parent log execution or the workflow execution (in case parent is None)