<feature>Map (parallel flow) can execute an outer subworkflow in IPython engines</feature>
Executing a subworkflow in IPython engines was not working because the engines could not locate the file that describes it. And if the engines were located in different machines, the user would have to add the subworkflow on each machine. To avoid that, if a subworkflow needs to be executed in the engines, it is first transformed into a Group, and then sent to the engines. However, note that this only applies to an outer subworkflow. If the main module being executed in the engines is a Subworkflow / Group module with subworkflows inside, these inner subworkflows are not transformed. In this case, we would need to recursively check for subworkflows and change the inner pipelines, and I don't know if we want to do that.

vistrails/packages/parallelflow/map.py:
- Map.updateFunctionPort: added check for subworkflow and transformation from subworkflow to group