<feature>Added information about the machine from each IPython engine</feature>
Each execution from an engine comes with machine information. Then, machines are added to the local log, ids are assigned to them and module / group executions (coming from IPython engines) are associated with the right machine information. No changes in the current schema were necessary. Note that, even if the engines are placed in the same machine, each one is assigned to a different Machine object, which allows flexibility when engines are placed in different machines.

vistrails/packages/parallelflow/map.py:
- execute_wf: the function executed in the engines now outputs the machine information
- Map.updateFunctionPort: machine information is added for each engine in the log