<feature>IPython controller and engines are now started from VisTrails</feature>
IPython controller and engines can now be started inside VisTrails. Before, users needed to do that by using the IPython command-line tool. Now, a new menu was inserted for the package (under the Packages menu) that allows users to start controller and engines, restart controller and engines, and also stop the controller (which automatically stops the engines). The code does not handle the controller and the engines by using the command-line tool; this is done programmatically. Note that, for now, using the menu only allows users to deal with LOCAL controller and LOCAL engines.
In a minor modification, the Map module is now not cacheable.

vistrails/packages/parallelflow/map.py:
- Map: the class now inherits the NotCacheable class

vistrails/packages/parallelflow/init.py:
- menu_items(): function to add the menu items, under the Packages menu, for the parallel flow package
- start_local_controller(), add_local_engines(), restart_controller(), restart_engines(), stop_controller(): callback functions for the menu items

vistrails/packages/parallelflow/ipython_set.py:
- IPythonSet: class that defines an IPython set, which contains a controller and multiples engines
- QWarningDialog: a warning dialog to interact with users with respect to the controller and the engines
- QAddEnginesDialog: a dialog to get the number of engines that the users want to add