<bugfix>Execution of workflows in IPython engines</bugfix> <feature>User does not need to create IPython controller manually anymore</feature>
When trying to re-execute a workflow with different parameter values in IPython engines, the expected output values were not being returned. The issue was solved by erasing the cache in the engines before a new execution.

Also, before, the user had to start the controller manually. As there can be only one type of controller, local controller, it does not make much sense for the user to start it manually. Now, the controller is automatically started when the user adds IPython engines.

vistrails/packages/parallelflow/init.py
- menu_items: removed menu entries to start and restart controller
- add_engines: function now also starts the controller

vistrails/packages/parallelflow/map.py
- execute_wf: importing the interpreter and erasing the cache now