<feature>Added saving/loading SubWorkflows to/from the DB as part of a vistrail bundle</feature>
core/vistrail/controller.py:
- VistrailController.parse_abstraction_name: *moved to core/sub_module.py
Also updated all the calls to reflect this
- VistrailController.check_abstraction:
An abstraction was not being matched if it did not contain the exact namespace, now it searches all with the same name by checking the (abs_name, '') key.
- VistrailController.ensure_abstractions_loaded:
store all abstractions with a given name as a list with the (name, '') key.

core/modules/sub_module.py:
- read_vistrail_from_db: *new - reads a vistrail from db and converts to Vistrail
- parse_abstraction_name: *new - moved from controller so that it can be used without the controller

db/services/io.py:
- get_db_object_list: Skip abstractions when listing vistrails
- get_db_abstraction_modification_time: *new
Returns modification time for abstraction or None if not found
- get_db_abstraction_ids_from_vistrail: *new
Returns list of ids in database for a vistrail
- open_vistrail_bundle_from_db:
Added open abstractions
- save_vistrail_bundle_to_db:
Added save/update abstractions
- save_abstractions_to_db: *new
Saves list of abstractions to the db as a Vistrail with special properties