<feature>Add port spec translation from schema 1.0.3 to 1.0.2</feature>
Tommy's changes to terminator.vt materialized an issue where new vistrails could not be opened in older versions of VisTrails.  This was due to an absence of code to transform port spec objects in the v1.0.3 schema to those in the v1.0.2 schema.  This has now been added.

Note: I would argue that because schema 1.0.3 has not been finalized, all examples should be in the 1.0.2 format.  We do not want a case where an intermediate v1.0.3 file is one of our examples.

db/versions/v1_0_2/translate/v1_0_3.py:
  - translateVistrail: translate port specs
  - translateWorkflow: translate port specs
  - translateRegistry: translate port specs