<bugfix>Fix issues with package reloading</bugfix> <feature>Add support for customizing which packages are (not) reloaded</feature> <ticket>#646</ticket>
This commit addresses a few issues with package reloading.  First, it takes a slightly different method of identifying the names of the packages added so that it captures the dependency list better.  This fixes the issue Colin Talbert ran into with the SAHM package.  Second, it adds an unload call to remove_package which means that dependencies are unloaded when a package is.  This fixes an issue where package developers who fixed an error in the _modules list or _input_ports would not see a change upon reloading.  Finally, it incorporates Remi's ideas from commit 22b065b9e35b9fc54c87a8e87120524bd3e7de08 on the java branch, adding support for customizing which packages are or are not reloaded.

core/modules/module_registry.py:
  - ModuleRegistry.remove_package: add call to unload so that the package's init.py is unloaded when reloading a package

core/modules/package.py:
  - Package._override_import: add more options for forcing package loads/unloads
  - Package._import: add util methods to help determine when a package should be added to the dependency list, crawl the module hierarchy to identify all loaded packages
  - Package.initialize: pull in the options to override package loads/unloads