WIP: matplotlib package updates with more tools for semi-automated generation
This is still in progress, but some of the plots are functional and much of the infrastructure is in place to more easily generate the rest.  See the README if you are interested in how things work.  Note that it is not recommended to use this for non-development work at the moment.