Fixes StandardWidgetSheetTab#setDimension()
It previously changed the row and column spinbox but because their
editingFinished() was connected instead of valueChanged(), this didn't
trigger an actual change in the dimensions of the sheet.

valueChanged() cannot be used because it would be fired while the user
types and could provoke data loss (going from size 8 to size 10 means
that it is first resized to 1 before the user presses '0').