Example DAT code for the URL package
This is really more of an example than an actual useful package.

Adds an ImageFromWeb plot that shows an image from a URL variable.

Adds two url variable loaders: one creates a variable from direct input,
the other loads a .url file.

Also provides a conversion from a number to a URL.