Changes hooking method for the spreadsheet
Instead of passing flags and specific keyword argulents to the
SpreadsheetWindow's constructor and propagating them around, we set a
global dictionary on the SpreadsheetController that is used throughout
the package. This allows more flexibility, fixes some issues and use an
approach more similar to the way this was implemented in the main
VisTrails application.