<feature>Support package identifier changes.</feature>
In the past, it was not possible to assign more than one identifier to a package or provide any path for migrating to a new identifier. This commit adds the ability to specify a list, old_identifiers, of identifiers that were used for a package in the past.  The current identifier should always be listed in the identifier field.  In addition, you should change the version when making an identifier switch.

vistrails/core/modules/module_registry.py:
  - ModuleRegistry.setup_indices: add entries in self.packages for each identifier, whether it be in identifier or old_identifiers
  - ModuleRegistry.add_package: add entries in self.packages for each identifier, whether it be in identifier or old_identifiers

vistrails/core/modules/package.py:
  - Package.set_defaults: init old_identifiers
  - Package.set_properties: set old_identifiers if specified

vistrails/core/packagemanager.py:
  - PackageManager.identifier_is_available: check old_identifiers

vistrails/core/upgradeworkflow.py:
  - UpgradeWorkflowHandler.check_port_spec: fix old identifiers in port spec strings

vistrails/gui/preferences.py:
  - QPackagesWidget.package_removed: call populate_lists, don't rely on cache since packages with user-specified prefixes that are late-disabled won't be there.