WIP: Improve storage bundles to be uniform across storage types, improve abstraction.
The goal here is to improve the current SaveBundle so that it stores any object, has a manifest to make serialization more straightforward and delegates serialization of individual objects to specific, storage-specific classes.  Also introduces a DirectoryBundle and DirectoryLocator to, among other things, support auto-save of full bundles.  In addition, we envision a more generic bundle to be customizable so users may choose to include, for example, data or packages in the .vt format.

db/services/bundle.py:
  - BundleObj: new class
  - BundleObjDictionary: new class
  - Bundle: new class
  - Serializer: new class
  - FileSerializer: new class
  - ThumbnailFileSerializer: new class
  - AbstractionFileSerializer: new class
  - XMLFileSerializer: new class
  - RegistryXMLSerializer: new class
  - LogXMLSerializer: new class
  - Manifest: new class
  - FileManifest: new class
  - DirectoryBundle: new class
  - ZIPBundle: new class
  - DBManifest: new class
  - DBBundle: new class
  - VistrailBundle: new class
  - WorkflowBundle: new class
  - LogBundle: new class
  - RegistryBundle: new class
  - MySQLDatabase: new class
  - SQLite3Database: new class
  - TestBundles: new class

db/services/locator.py:
  - BaseLocator.from_url: add DirectoryLocator
  - DirectoryLocator: new class